requirements.txt
# System
pyinstaller
openai
boto3
anthropic
google-genai
tiktoken
numpy
openpyxl
PyQt5
PyQt-Fluent-Widgets[full]~=1.10.0
opencc
beautifulsoup4
lxml
chardet
PyYAML
requests
httpx[http2]
curl_cffi

# Tools
python-pptx
polib
pandas
rich
tqdm
jaconv
python-rapidjson
protobuf==4.25.7
# 获取iso语言代码对应的友好名称
language-data
langcodes
msgspec

# mediapipe
absl-py
attrs>=19.1.0
flatbuffers>=2.0
matplotlib
sounddevice>=0.4.4

# BabelDOC
# onnx 新版本打包后不能运行，这里使用 onnx==1.16.1 onnxruntime==1.20.1
# rapidocr-onnxruntime 依赖了 opencv-python，所以去除了 opencv-python-headless
# 不使用 babeldoc.document_il.midend.detect_scanned_file.DetectScannedFile 所以去除 scikit-image>=0.25.2
bitstring>=4.3.0
configargparse>=1.7
httpx[socks]>=0.27.0
huggingface-hub>=0.27.0
numpy>=2.0.2
onnx==1.16.1
onnxruntime==1.20.1
openai>=1.59.3
orjson>=3.10.14
charset-normalizer>=2.0.0
cryptography>=36.0.0
peewee>=3.17.8
psutil>=7.0.0
pymupdf==1.26.7
rich>=13.9.4
toml>=0.10.2
tqdm>=4.67.1
xsdata[cli,lxml,soap]>=24.12
msgpack>=1.1.0
pydantic>=2.10.6
tenacity>=9.0.0
freetype-py>=2.5.1
tiktoken>=0.9.0
levenshtein>=0.27.1
rapidocr-onnxruntime>=1.4.4
pyzstd
rtree
uharfbuzz>=0.50.2
scikit-learn>=1.7.1
